requirements.txt
django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1

# ML / Data
scikit-learn==1.4.0
xgboost==2.0.3
pandas==2.2.0
numpy==1.26.3
joblib==1.3.2

# Maps / Geo
folium==0.15.1
geopy==2.4.1

# Visualization
plotly==5.18.0

# Utils
python-dotenv==1.0.0
Pillow==10.2.0
